cogl: Start adding support for emscripten/html5
This makes lots of emscripten portability updates and support
throughout cogl

Note: Instead of adding emscripten to our autoconf/automake build
scripts it's more important that native builds of Rig can be distributed
with source and an application built with Rig can be compiled and
packaged for deployment as a html5 application.

For now this starts by adding a vanilla gnu Makefile
(Makefile.emscripten) that could be installed and used by tools to
compile cogl for emscripten.