Start adding support for emscripten/html5
This makes lots of emscripten related changes throughout rig

Note: instead of adding emscripten to our autoconf/automake build
scripts it's more important that native builds of Rig can be distributed
with source and an application built with Rig can be compiled and
packaged for deployment as a html5 application.

This adds vanilla gnu Makefiles (Makefile.emscripten) that can be
installed and used by tools to build Rig.

This adds an initial rig-package-html5 helper script

This script can be distributed with Rig so that developers can compile
and package an application for distribution as a html5 application. This
is simply a prototype script for now. Eventually it should be possible
to control packaging from the Rig ui and any text based tool should
be updated to be more portable.

A lot of changes were made to remove play mode + make the simulator
responsible for ui loading to help minimize the responsibility of
a web based frontend:

This work is aiming to remove some of the complexity around the
simulator and frontend having to track whether it is currently in play
or edit mode and instead we will aim for that concept to be internal to
the editor, probably by making the editor manage separate
frontend+simulator pairs for edit and play mode and it will composite
the rendering from those nested frontends via its own frontend.

To be better suited to web based usage models with a web client acting
as a frontend this allows the simulator to be responsible for loading a
.rig file and defining canonical object IDs.  This removed the need for
allocating temporary IDs while we don't have a need for the frontend to
create its own objects.